Performance improvement (e.g. Removed unnecessary runnable), deallocated code. Now doesn't rely on World#getName() to assume that the player is on session. Player speed is now accurate and dependent on PlayerMoveEvent.

+++ src/main/java/com/github/thethingyee/bridgingpractice/listeners/ChangeWorld.java
@@ -1,7 +1,6 @@
 package com.github.thethingyee.bridgingpractice.listeners;
 
 import com.github.thethingyee.bridgingpractice.BridgingPractice;
-import com.github.thethingyee.bridgingpractice.PlayerSpeed;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerChangedWorldEvent;
@@ -16,13 +15,13 @@ public class ChangeWorld implements Listener {
 
     @EventHandler
     public void onChangeWorld(PlayerChangedWorldEvent e) {
-        PlayerSpeed playerSpeed = bridgingPractice.getPlayerSpeed();
         if(!bridgingPractice.getActiveSessions().containsKey(e.getPlayer())) return;
         if(e.getPlayer().getWorld().equals(bridgingPractice.getActiveSessions().get(e.getPlayer()).getAssignedWorld())) {
-            playerSpeed.stopPlayerSpeed(e.getPlayer());
             bridgingPractice.getActiveSessions().get(e.getPlayer()).setSchematicName(null);
-            bridgingPractice.getActiveSessions().get(e.getPlayer()).getScoreboard().delete();
-            bridgingPractice.getActiveSessions().get(e.getPlayer()).setScoreboard(null);
+            if(bridgingPractice.getActiveSessions().get(e.getPlayer()).getScoreboard() != null) {
+                bridgingPractice.getActiveSessions().get(e.getPlayer()).getScoreboard().delete();
+                bridgingPractice.getActiveSessions().get(e.getPlayer()).setScoreboard(null);
+            }
         }
     }
 }
