+++ src/main/java/com/github/thethingyee/bridgingpractice/listeners/PlayerInteract.java
@@ -4,15 +4,10 @@ import com.github.thethingyee.bridgingpractice.BridgingPractice;
 import com.github.thethingyee.bridgingpractice.GUIManager;
 import com.github.thethingyee.bridgingpractice.utils.Offsets;
 import com.github.thethingyee.bridgingpractice.utils.Session;
-import org.apache.commons.io.FileUtils;
 import org.bukkit.*;
-import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerInteractEvent;
-
-import java.io.File;
-import java.io.IOException;
 
 public class PlayerInteract implements Listener {
 
@@ -33,14 +28,17 @@ public class PlayerInteract implements Listener {
         GUIManager guiManager = bridgingPractice.getGuiManager();
         if(!bridgingPractice.getActiveSessions().containsKey(e.getPlayer())) return;
         Session currentSession = bridgingPractice.getActiveSessions().get(e.getPlayer());
+
         if (e.getPlayer().getWorld().equals(bridgingPractice.getActiveSessions().get(e.getPlayer()).getAssignedWorld())) {
-            if(e.getPlayer().getItemInHand().getItemMeta() == null) return;
+            if (e.getPlayer().getItemInHand().getItemMeta() == null) return;
             if (e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getChangeColMeta())) {
                 guiManager.changeColorGUI(e.getPlayer());
                 e.getPlayer().sendMessage(bridgingPractice.prefix + ChatColor.GREEN + "You have opened the Change Color GUI.");
-            } else if (e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getRestartMeta())) {
+                return;
+            }
+            if (e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getRestartMeta())) {
                 double[] offset = Offsets.getOffsets(currentSession.getSchematicName());
-                e.getPlayer().teleport(new Location(Bukkit.getWorld(e.getPlayer().getUniqueId().toString().replaceAll("-", "")), 0.0, 128.0, 0.0).add(offset[0], offset[1]+1, offset[2]));
+                e.getPlayer().teleport(new Location(currentSession.getAssignedWorld(), 0.0, 128.0, 0.0).add(offset[0], offset[1] + 1, offset[2]));
                 if (currentSession.getBlockPlaced() != null) {
                     for (Location loc : currentSession.getBlockPlaced()) {
                         e.getPlayer().getWorld().getBlockAt(loc).setType(Material.AIR);
@@ -48,27 +46,34 @@ public class PlayerInteract implements Listener {
                 }
                 currentSession.setBlocksPlaced(0);
                 e.getPlayer().sendMessage(bridgingPractice.prefix + ChatColor.GREEN + "Successfully restarted world.");
-            } else if (e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getLeaveMeta())) {
+                return;
+            }
+            if (e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getLeaveMeta())) {
                 if (currentSession.getAssignedWorld() != null) {
-                    bridgingPractice.getActiveSessions().get(e.getPlayer()).leaveSession(bridgingPractice, e.getPlayer());
+                    if (!bridgingPractice.getActiveSessions().get(e.getPlayer()).leaveSession(bridgingPractice, e.getPlayer())) {
+                        e.getPlayer().sendMessage(bridgingPractice.prefix + ChatColor.RED + "Can't leave session.");
+                        return;
+                    }
+                    e.getPlayer().sendMessage(bridgingPractice.prefix + ChatColor.GREEN + "Successfully left session...");
                 }
                 e.getPlayer().getInventory().clear();
-            } else if(e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getRefillMeta())) {
+            } else if (e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getRefillMeta())) {
                 e.getPlayer().getInventory().clear();
                 guiManager.giveInventoryItems(e.getPlayer(), currentSession.getWoolColor());
+                return;
             }
+            return;
         }
-        if(currentSession.isSpectatingPlayer()) {
-            if(currentSession.getSpectating().getWorld() == e.getPlayer().getWorld()) {
-                if(e.getPlayer().getItemInHand().getItemMeta() == null) return;
-                if(e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getSpectatorLeave())) {
-                    currentSession.getScoreboard().delete();
-                    currentSession.setScoreboard(null);
-                    currentSession.setSpectating(null);
-                    e.getPlayer().teleport(Bukkit.getWorld(bridgingPractice.getConfig().getString("defaults.world")).getSpawnLocation());
-                    e.getPlayer().sendMessage(bridgingPractice.prefix + ChatColor.GREEN + "Teleported!");
-                    e.getPlayer().getInventory().clear();
-                }
+
+        if(currentSession.isSpectatingPlayer() && currentSession.getSpectating().getWorld() == e.getPlayer().getWorld()) {
+            if(e.getPlayer().getItemInHand().getItemMeta() == null) return;
+            if(e.getPlayer().getItemInHand().getItemMeta().equals(guiManager.getSpectatorLeave())) {
+                currentSession.getScoreboard().delete();
+                currentSession.setScoreboard(null);
+                currentSession.setSpectating(null);
+                e.getPlayer().teleport(Bukkit.getWorld(bridgingPractice.getConfig().getString("defaults.world")).getSpawnLocation());
+                e.getPlayer().sendMessage(bridgingPractice.prefix + ChatColor.GREEN + "Teleported!");
+                e.getPlayer().getInventory().clear();
             }
         }
     }
